added new codes from NajmiddinG

+++ base/views.py
@@ -1,6 +1,6 @@
 from django.contrib.auth import authenticate, login
 from django.contrib.auth import logout
-from rest_framework import status, generics
+from rest_framework import status, generics, viewsets
 
 from rest_framework.views import APIView
 from rest_framework.response import Response
@@ -11,7 +11,6 @@ from rest_framework.permissions import IsAuthenticated
 from .models import Mahsulot_filter, Mahsulot, Oluvchi, Savdo, Chiqim
 from .serializers import MahsulotFilterSerializer, MahsulotSerializer, \
     OluvchiSerializer, SavdoSerializer,ChiqimSerializer
-
 
 
 class LoginAPIView(APIView):
@@ -26,60 +25,35 @@ class LoginAPIView(APIView):
         else:
             return Response({'message': 'Invalid credentials'}, status=status.HTTP_401_UNAUTHORIZED)
 
+
 class LogoutAPIView(APIView):
     permission_classes = [IsAuthenticated]
 
     def post(self, request):
-        # Get the user's token
         token = request.auth
-
-        # Delete the user's token
-        if token:
-            Token.objects.filter(key=token).delete()
-
-        # Perform the logout
+        if token: Token.objects.filter(key=token).delete()
         logout(request)
 
         return Response("Logged out successfully", status=status.HTTP_200_OK)
 
+
 # mahsulot filter
-class MahsulotFilterlarAPIView(generics.ListCreateAPIView):
+class MahsulotFilterAPIView(viewsets.ModelViewSet):
     permission_classes = [IsAuthenticated]
     queryset = Mahsulot_filter.objects.all()
     serializer_class = MahsulotFilterSerializer
-
-
-class MahsulotFilterAPIView(generics.RetrieveUpdateDestroyAPIView):
-    permission_classes = [IsAuthenticated]
-    queryset = Mahsulot_filter.objects.all()
-    serializer_class = MahsulotFilterSerializer
-
-
 
 
 #mahsulot
-class MahsulotlarAPIView(generics.ListCreateAPIView):
-    permission_classes = [IsAuthenticated]
-    queryset = Mahsulot.objects.all()
+class MahsulotAPIView(viewsets.ModelViewSet):
     serializer_class = MahsulotSerializer
-
-
-
-class MahsulotAPIView(generics.RetrieveUpdateDestroyAPIView):
-    permission_classes = [IsAuthenticated]
     queryset = Mahsulot.objects.all()
-    serializer_class = MahsulotSerializer
+    permission_classes = [IsAuthenticated]
 
 
 
 #oluvchi uchun ipis
-class OluvchilarAPIView(generics.ListCreateAPIView):
-    permission_classes = [IsAuthenticated]
-    queryset = Oluvchi.objects.all()
-    serializer_class = OluvchiSerializer
-
-
-class OluvchiAPIView(generics.RetrieveUpdateDestroyAPIView):
+class OluvchiAPIView(viewsets.ModelViewSet):
     permission_classes = [IsAuthenticated]
     queryset = Oluvchi.objects.all()
     serializer_class = OluvchiSerializer
@@ -87,26 +61,14 @@ class OluvchiAPIView(generics.RetrieveUpdateDestroyAPIView):
 
 
 #for savdo
-class SavdolarAPIView(generics.ListCreateAPIView):
-    permission_classes = [IsAuthenticated]
-    queryset = Savdo.objects.all()
-    serializer_class = SavdoSerializer
-
-
-class SavdoAPIView(generics.RetrieveUpdateDestroyAPIView):
+class SavdoAPIView(viewsets.ModelViewSet):
     permission_classes = [IsAuthenticated]
     queryset = Savdo.objects.all()
     serializer_class = SavdoSerializer
 
 
 # for chiqim
-class ChiqimlarAPIView(generics.ListCreateAPIView):
-    permission_classes = [IsAuthenticated]
-    queryset = Chiqim.objects.all()
-    serializer_class = ChiqimSerializer
-
-
-class ChiqimAPIView(generics.RetrieveUpdateDestroyAPIView):
+class ChiqimAPIView(viewsets.ModelViewSet):
     permission_classes = [IsAuthenticated]
     queryset = Chiqim.objects.all()
     serializer_class = ChiqimSerializer
